course-4/Программирование/Microservices/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
matplotlib>=3.7.0

